LayoutTests/fast/regions/auto-height/first-region-auto-height-break-after-always-expected.txt

Test for regions auto height prototype.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS h2RegionList.item(0).id is "region"
PASS p1RegionList.length is 1
PASS p1RegionList.item(0).id is "region2"
PASS successfullyParsed is true

TEST COMPLETE
